requirements.txt
aiogram>=3.24.0
aiosqlite
sqlalchemy[asyncio]
environs
PyYAML
loguru
